test: suíte de testes, CI e 2 bugs de classificação que ela encontrou
318 testes cobrindo o que a ferramenta tem de mais frágil: o parsing.
normalize_mac, is_random_mac, TTL, tabela ARP (ip neigh e arp -a), saída
do nmap, NBSTAT, XML do UPnP, validação de URL de aparelho, nomes,
fabricantes, histórico, achados e a matriz de classificação.

Dois bugs que a matriz expôs, ambos de casamento por substring:

- "Intelbras" casava com "intel" e virava COMPUTADOR. A Intelbras é uma
  das marcas mais comuns numa rede doméstica brasileira, e vende câmera
  e roteador — não computador. Fabricante agora casa por palavra.

- "Sony Interactive Entertainment" (PlayStation) casava com a entrada de
  TV, que vinha antes da de console na lista. Reordenado.

Na mesma família, prevenidos: um switch de rede chamado "switch-sala"
virava CONSOLE (removido "switch" solto — "nintendo" e o OUI já cobrem o
console), e um banner com "watchdog" virava WEARABLE.

Infra:
- CI: pytest em 3.9–3.13 no Linux, mais macOS, ruff, shellcheck e
  verificação de empacotamento.
- requires-python sobe para 3.9 (3.8 é EOL desde out/2024 e nunca foi
  testado); a versão passa a ter fonte única em alive/__init__.py.
- Metadados de licença na forma SPDX: a tabela e o classifier param de
  ser aceitos em fev/2027. Build agora sai sem nenhum aviso.
- install.sh: removida a variável BLUE, definida e nunca usada (SC2034).

+++ alive/scanner.py
@@ -65,7 +65,7 @@ def _ping_cmd(ip: str, timeout: float) -> list[str]:
         # -W em milissegundos no macOS
         return ["ping", "-c", "1", "-t", "1", "-W", str(int(timeout * 1000)), ip]
     # Linux: -W em segundos (inteiro, mínimo 1)
-    return ["ping", "-c", "1", "-w", str(max(1, int(round(timeout)))), ip]
+    return ["ping", "-c", "1", "-w", str(max(1, round(timeout))), ip]
 
 
 def _ping_one(ip: str, timeout: float) -> Optional[dict]:
